Fix various issues in flush workflow and when inserting single elements
1) The flush workflow has been fixed in order to account for differences
in the way MGPU and our custom algorithms work and expect data to be passed.
2) Furthermore now the flush workflow can handle a flush with no inserts
without crashing.
3) The insert of a single element in SparseGridGpu was allocating two
different blocks for data and mask, causing all sorts of problems. The
issue is now solved and insert works properly.
4) The solve_conflicts operation in the BlockFunctor now properly updates
also the keys of the internal data structure to be consistent with the
segreduced data.