Fix to allow indexing of blocks in terms of 3D coordinates
Previously the block id used, e.g., to determine the pool to use for
the current block was based only on blockIdx.x, therefore failing if
the kernel was called with a 2D or 3D layout of blocks.
Now blocks are indexed in terms of the dim3 object blockIdx, therefore
supporting also 2D and 3D grids.