Add SparseGridGpu as geometry on top of linear BlockMapGpu structure
SparseGridGpu is now a specialization of BlockMapGpu adding a geometry
on top of the linear BlockMapGpu structure.
The client can specify the geometric properties of the SparseGridGpu at
construction (e.g. dim, dimension sizes, ...) and then address elements
using dim-dimensional coordinates.
The linearization is done in a way that ensures all-dimensional locality
of a data block.
The current status of SparseGridGpu is that the geometry is handled properly
but the stencil computation facilities are still missing/untested.